requirements.txt
Django==3.0.3
gunicorn==20.0.4
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2==2.8.4
whitenoise==5.0.1
python-dotenv==0.11.0
